corrigiendo el main

+++ Source.cpp
@@ -10,9 +10,9 @@ void main()
 	cin>>nx;
 	cout<<"ingresar ny"<<endl;
 	cin>>ny;
-	for(int i=0;i<nx;i++)
+	for(float i=0;i<nx;i++)
 	{
-		for(int j=0;j<ny;j++)
+		for(float j=0;j<ny;j++)
 		{
 			dir=a.calcular_distan(i,j,nx,ny);
 			dir.mostrar();
